requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
numpy>=1.21.0
pillow>=8.3.1
requests>=2.26.0 